benchmarks/locobench_agent/tasks/python_game_engine_expert_032_architectural_understanding_expert_01/environment/project/tests/test_suite.py
```python
"""
LedgerQuest Engine – Test-Suite
===============================

Integration-style and unit-style tests that exercise the contract of the
LedgerQuest Engine’s most critical building blocks: physics, AI behaviour, the
event bus, and persistence.  100 % real AWS resources are *not* required—the
`moto` library is used to spin up in-memory mocks for DynamoDB, S3, and
EventBridge so that the tests run on any CI agent with no credentials.

The tests are intentionally opinionated: they codify the rules that keep the
game-engine deterministic (conservation of energy in the physics module),
predictable (finite-state AI behaviour), and compliant (audit events are
persisted and retrievable).

To execute:
    $ pip install pytest moto boto3 hypothesis
    $ pytest -q LedgerQuestEngine/tests/test_suite.py
"""
from __future__ import annotations

import json
import os
import time
import uuid
from decimal import Decimal
from typing import Dict, Generator, List, Tuple

import boto3
import pytest
from botocore.exceptions import ClientError
from hypothesis import given, settings
from hypothesis import strategies as st
from moto import mock_dynamodb, mock_events, mock_s3

# ------------------------------------------------------------------------------
# Optional real import ‑- alongside an internal stub so this test-suite remains
# runnable even without the full LedgerQuest code-base present.
# ------------------------------------------------------------------------------

try:
    # Uncomment in real project:
    # from ledgerquest_engine import physics, ai, event_bus, persistence
    raise ImportError  # noqa: ERA001 – Delete once real package is available.
except ImportError:  # pragma: no cover
    # ------------------------------------------------------------------
    # ░░░  S T U B S   (these shadow the public contract only)  ░░░
    # ------------------------------------------------------------------
    class physics:  # noqa: D101 – Stub
        GRAVITY = 9.81

        @staticmethod
        def kinetic_energy(mass: float, velocity: float) -> float:
            return 0.5 * mass * velocity**2

        @staticmethod
        def potential_energy(mass: float, height: float) -> float:
            return mass * physics.GRAVITY * height

        @staticmethod
        def total_energy(mass: float, velocity: float, height: float) -> float:
            return physics.kinetic_energy(mass, velocity) + physics.potential_energy(
                mass, height
            )

        @staticmethod
        def update_velocity(velocity: float, dt: float) -> float:
            """Naïve Euler integration in constant gravity field."""
            return velocity - physics.GRAVITY * dt

    class ai:  # noqa: D101 – Stub
        class BehaviourTree:  # noqa: D101 – Stub
            """Single-state AI for demonstration only."""

            def decide(self, state: Dict[str, bool]) -> str:
                if state.get("enemy_visible"):
                    return "attack"
                if state.get("low_health"):
                    return "retreat"
                return "patrol"

    class event_bus:  # noqa: D101 – Stub
        class EventBus:  # noqa: D101 – Stub
            def __init__(self) -> None:
                self._events: List[Dict] = []

            # Public API of the real EventBridge wrapper
            def publish(self, event: Dict) -> None:
                event["id"] = str(uuid.uuid4())
                self._events.append(event)

            def last_event(self) -> Dict | None:
                return self._events[-1] if self._events else None

    class persistence:  # noqa: D101 – Stub
        class DynamoPersistence:  # noqa: D101 – Stub
            def __init__(self, table) -> None:
                self._table = table

            def save_state(self, pk: str, data: Dict) -> None:
                self._table.put_item(Item={"pk": pk, "state": json.dumps(data)})

            def load_state(self, pk: str) -> Dict | None:
                try:
                    res = self._table.get_item(Key={"pk": pk})
                except ClientError:
                    return None
                item = res.get("Item")
                return json.loads(item["state"]) if item else None


# ------------------------------------------------------------------------------
#                              F I X T U R E S
# ------------------------------------------------------------------------------


@pytest.fixture(scope="session", autouse=True)
def _aws_env() -> None:
    """
    Ensure boto3 sees *some* AWS credentials (even fake) so that the client
    constructor does not bail out early in CI containers.
    """
    os.environ.setdefault("AWS_ACCESS_KEY_ID", "testing")
    os.environ.setdefault("AWS_SECRET_ACCESS_KEY", "testing")
    os.environ.setdefault("AWS_DEFAULT_REGION", "us-east-1")


@pytest.fixture(scope="function")
def dynamodb_table(_aws_env) -> Generator:
    with mock_dynamodb():
        resource = boto3.resource("dynamodb", region_name="us-east-1")
        table = resource.create_table(
            TableName="game-state",
            KeySchema=[{"AttributeName": "pk", "KeyType": "HASH"}],
            AttributeDefinitions=[{"AttributeName": "pk", "AttributeType": "S"}],
            BillingMode="PAY_PER_REQUEST",
        )
        table.wait_until_exists()
        yield table


@pytest.fixture(scope="function")
def s3_bucket(_aws_env) -> Generator:
    with mock_s3():
        resource = boto3.resource("s3", region_name="us-east-1")
        bucket_name = f"assets-{uuid.uuid4()}"
        bucket = resource.Bucket(bucket_name)
        bucket.create()
        yield bucket


@pytest.fixture(scope="function")
def eventbridge_bus(_aws_env) -> Generator:
    with mock_events():
        client = boto3.client("events", region_name="us-east-1")
        bus_name = f"ledgerquest-bus-{uuid.uuid4()}"
        client.create_event_bus(Name=bus_name)
        yield client, bus_name


# ------------------------------------------------------------------------------
#                            P H Y S I C S   T E S T S
# ------------------------------------------------------------------------------


class TestPhysicsEnergy:
    mass = 10.0
    initial_height = 100.0
    dt = 0.01
    steps = int(100 / dt)  # Simulate ~100 s of free-fall

    def test_energy_conservation(self) -> None:
        """
        The (potential + kinetic) energy of a closed system must remain constant.
        We allow ≤ 0.5 % numerical drift to accommodate the simple Euler solver.
        """
        velocity: float = 0.0
        height: float = self.initial_height
        init_energy = physics.total_energy(
            self.mass,
            velocity,
            height,
        )

        t = 0.0
        while height > 0:
            velocity = physics.update_velocity(velocity, self.dt)
            height = max(0.0, height + velocity * self.dt)
            t += self.dt

        final_energy = physics.total_energy(
            self.mass,
            velocity,
            height,
        )
        drift_pct = abs(init_energy - final_energy) / init_energy * 100
        assert (
            drift_pct <= 0.5
        ), f"Energy drift exceeded tolerance: {drift_pct:.4f} % after {t:.2f}s"

    @given(
        mass=st.floats(min_value=0.5, max_value=100, allow_nan=False, allow_infinity=False),
        velocity=st.floats(min_value=-50, max_value=50, allow_nan=False, allow_infinity=False),
    )
    @settings(deadline=1000)
    def test_kinetic_energy_symmetry(self, mass: float, velocity: float) -> None:
        """
        Property-based test: Eₖ should be the same for +v and −v.
        """
        assert physics.kinetic_energy(mass, velocity) == pytest.approx(
            physics.kinetic_energy(mass, -velocity)
        )


# ------------------------------------------------------------------------------
#                           A I   B E H A V I O U R   T E S T S
# ------------------------------------------------------------------------------


@pytest.mark.parametrize(
    "state,expected",
    [
        (dict(enemy_visible=True, low_health=False), "attack"),
        (dict(enemy_visible=False, low_health=True), "retreat"),
        (dict(enemy_visible=False, low_health=False), "patrol"),
    ],
)
def test_behaviour_tree_decision(state: Dict[str, bool], expected: str) -> None:
    """Finite-state decisions must respect the priority order defined above."""
    tree = ai.BehaviourTree()
    decision = tree.decide(state)
    assert decision == expected


# ------------------------------------------------------------------------------
#                         E V E N T   B U S   T E S T S
# ------------------------------------------------------------------------------


def test_event_published_on_asset_upload(
    s3_bucket, eventbridge_bus
) -> None:  # noqa: D401
    """
    GIVEN an asset (e.g. GLTF model) uploaded to S3,
    WHEN the engine's object-store lambda runs,
    THEN an EventBridge message must be emitted with contextual metadata.
    """
    client, bus_name = eventbridge_bus
    bus = event_bus.EventBus()

    # --- Step 1: Simulate an asset upload
    key = "models/dragon.gltf"
    file_contents = b"pretend-this-is-a-valid-gltf-binary"
    s3_bucket.put_object(Key=key, Body=file_contents)

    # --- Step 2: Engine callback (simplified)
    event_payload = {
        "detail-type": "AssetUploaded",
        "source": "ledgerquest.engine.assets",
        "bucket": s3_bucket.name,
        "key": key,
        "size": len(file_contents),
        "timestamp": time.time(),
    }
    bus.publish(event_payload)

    # --- Step 3: Verify the local bus sees the event…
    last = bus.last_event()
    assert last is not None and last["detail-type"] == "AssetUploaded"

    # --- Step 4: …and marshal it to a mock EventBridge bus.
    response = client.put_events(
        Entries=[
            {
                "Source": last["source"],
                "DetailType": last["detail-type"],
                "Detail": json.dumps(last),
                "EventBusName": bus_name,
            }
        ]
    )
    assert response["FailedEntryCount"] == 0


# ------------------------------------------------------------------------------
#                         P E R S I S T E N C E   T E S T S
# ------------------------------------------------------------------------------


def test_save_and_load_game_state_roundtrip(dynamodb_table) -> None:
    """
    Game state persisted in DynamoDB must be lossless after JSON (de)serialise.
    """
    store = persistence.DynamoPersistence(dynamodb_table)

    pk = str(uuid.uuid4())
    game_state = {
        "player": {"x": Decimal("1.5"), "y": Decimal("3.2"), "hp": 95},
        "inventory": ["sword", "shield", "potion"],
        "timestamp": int(time.time()),
    }

    store.save_state(pk, game_state)
    loaded = store.load_state(pk)

    # Dynamo may coerce Decimals—ensure we get back what we put in.
    assert loaded == game_state
```