benchmarks/locobench_agent/tasks/python_game_engine_expert_032_architectural_understanding_expert_01/environment/project/requirements.txt
# ========================================================================
# LedgerQuest Engine – Production Dependency Manifest (requirements.txt)
# ========================================================================
# This file contains all Python-level dependencies (runtime + development)
# required to build, test, and deploy LedgerQuest Engine.  Version pins are
# locked to ensure deterministic CI/CD pipelines and to satisfy AWS Lambda
# container–image constraints.  Update via:
#
#     $ make deps-bump               # Runs `pip-compile` and integration tests
#
# Groups:
#   CORE ............... Engine runtime executed inside AWS Lambda/Fargate
#   SERVERLESS ......... AWS service SDKs & serverless-specific helpers
#   GAME/GRAPHICS ...... Real-time simulation libraries for local GPU workers
#   ECS / AI ........... Entity-Component-System, Behaviour Trees, path-finding
#   PERSISTENCE ........ Data-layer utilities (NoSQL, S3 streaming, etc.)
#   WEB/API ............ REST & WebSocket gateways (FastAPI for local tools)
#   OBSERVABILITY ...... Metrics, tracing, structured logging
#   DEV/TEST ........... Linting, formatting, static-analysis, test runners
# ========================================================================

###########
# CORE
###########
numpy==1.25.2
scipy==1.11.2
pydantic==1.10.12
dataclasses-json==0.5.9

###########
# SERVERLESS
###########
boto3==1.28.57               # AWS SDK for Python
botocore==1.31.57
aws-lambda-powertools==2.28.1
aws-xray-sdk==2.9.0
stepfunctions==2.0.0         # Data Science Step Functions SDK

###########
# GAME / GRAPHICS
###########
pygame==2.5.2                # Local physics & input harness
moderngl==5.8.2              # OpenGL 3.3+ wrapper for GPU workers
PyOpenGL==3.1.6
esper==2.3                   # ECS library
networkx==3.1                # Graph utilities (AI nav-mesh, quest trees)

###########
# ECS / AI
###########
numpy-quaternion==2022.4.2   # Quaternion math for 3D rotations
behave==1.2.6                # Simple behaviour-tree prototyping
scikit-learn==1.3.0          # ML helpers for adaptive difficulty

###########
# PERSISTENCE
###########
motor==3.3.1                 # Async MongoDB client (local debug)
sqlalchemy==2.0.20           # Audit event log sink
s3transfer==0.6.2

###########
# WEB / API
###########
fastapi==0.103.1
uvicorn[standard]==0.23.2
sse-starlette==1.4.0         # Server-sent events bridge for browser clients

###########
# OBSERVABILITY
###########
structlog==23.1.0
prometheus-client==0.17.1
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-exporter-otlp==1.19.0

###########
# DEV / TEST
###########
pytest==7.4.0
pytest-asyncio==0.21.1
coverage==7.2.7
black==23.7.0
isort==5.12.0
mypy==1.5.1
ruff==0.0.284
pre-commit==3.3.3
tox==4.6.4
boto3-stubs[essential]==1.28.57  # Typed AWS stubs for MyPy
types-requests==2.31.0.2
types-PyYAML==6.0.12.11
types-python-dateutil==2.8.19.14

###########
# NOTE
###########
# 1.   The AWS Lambda base image for Python 3.11 currently bundles specific
#      minor versions of `botocore`; ensure the pin above matches.  Use
#      `make validate-lambda-layer` before pushing to production.
#
# 2.   GPU-accelerated workers run inside AWS Fargate + ECS with NVIDIA
#      runtime; when upgrading `moderngl` or `PyOpenGL`, double-check the
#      CUDA compatibility matrix in docs/DEPLOY_GPU.md.
#
# 3.   All dev-helper dependencies are optional at runtime and may be
#      installed with:
#
#         $ pip install -r requirements.txt -r requirements-dev.txt
#
#      but we include them here so that a single “pip install -r” command
#      sets up a fully-featured local environment for new contributors.
# ========================================================================